refactor(widgets): Перенесла комопннеты Header, TextEditor, ImageUploadPreview в папку widgets
